+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Unreleased
 
+- FIXED: Record requests use the zone name instead of the zone ID (#7, #11)
 - CHANGED: Minimum Ruby version is now 3.2 (#9)
 - CHANGED: Moved CI from Travis to GitHub Actions (#9)
 - CHANGED: Minimum fog-core version is now 2.5 (#10)